usar a execução do jogo como uma função de avaliação
