+++ src/flappy_bird/neural_network/neural_network.py
@@ -2,8 +2,6 @@ from typing import Callable, List, Sequence, Tuple
 
 import numpy as np
 
-from flappy_bird.neural_network.evolution import Evolution
-
 
 def get_slice(sequence: Sequence, first_index: int, length: int):
     last_index = first_index + length
@@ -14,7 +12,7 @@ class PlayersPopulation:
     JUMP_THRESHOLD = 0.5
 
     def __init__(self, individuals_count_by_generation: int, neurons_disposition: Sequence[int],
-                 activation_functions: Sequence[Callable], use_backup: bool):
+                 activation_functions: Sequence[Callable], population):
         self.individuals_count_by_generation = individuals_count_by_generation
         self.neurons_disposition = neurons_disposition
         self.activation_functions = activation_functions
@@ -24,19 +22,16 @@ class PlayersPopulation:
         self.biases_lengths: List[int] = []
 
         for i in range(len(self.neurons_disposition) - 1):
-            size_layer_in = self.neurons_disposition[i]
+            size_layer_in = int(self.neurons_disposition[i])
             size_layer_out = self.neurons_disposition[i + 1]
 
             genes_count_by_individual += size_layer_in * size_layer_out + size_layer_out
             self.weights_shapes.append((size_layer_out, size_layer_in))
             self.biases_lengths.append(size_layer_out)
 
-        self.evolution = Evolution(genes_count_by_individual, self.individuals_count_by_generation,
-                                   mutation_probability=0.2, crossover_probability=0.2, use_backup=use_backup)
-
         self.weights: List[np.ndarray] = []
         self.biases: List[np.ndarray] = []
-        self.update_weights_and_biases_from_evolution()
+        self.update_weights_and_biases_from_evolution(population)
 
         self.output_layer: np.ndarray = None
 
@@ -66,11 +61,11 @@ class PlayersPopulation:
 
         return individual_weights, individual_biases
 
-    def update_weights_and_biases_from_evolution(self):
+    def update_weights_and_biases_from_evolution(self, population: list):
         weights: List[List[np.ndarray]] = [[] for _ in range(len(self.weights_shapes))]
         biases: List[List[np.ndarray]] = [[] for _ in range(len(self.biases_lengths))]
 
-        for individual in self.evolution.population:
+        for individual in population:
             arrays = self.deserialize_arrays_from_individual(individual)
             individual_weights, individual_biases = arrays
             for i in range(len(weights)):
@@ -85,7 +80,8 @@ class PlayersPopulation:
 
         for layer_weights, layer_biases, activation_function in zip(self.weights, self.biases,
                                                                     self.activation_functions):
-            result = activation_function(np.matmul(layer_weights, result) + layer_biases)
+            mul = np.matmul(layer_weights, result)
+            result = activation_function(mul + layer_biases)
         self.output_layer = np.reshape(result, (self.individuals_count_by_generation,))
 
     def should_jump(self, player_index: int):
