+++ src/main.py
@@ -2,7 +2,7 @@ import argparse
 
 import pygame
 
-from flappy_bird.gaming import Game
+from flappy_bird.evolution import Evolution
 
 parser = argparse.ArgumentParser(description="FlappyBird Genético.")
 parser.add_argument("--use-backup", action="store_true")
@@ -11,5 +11,6 @@ args = parser.parse_args()
 pygame.init()
 pygame.font.init()
 
-game = Game(450, 800, args.use_backup)
-game.run()
+evolution = Evolution(0.3, 0.3, args.use_backup, [2, 3, 1], 900)
+evolution.run_generation()
+
